WIP: package-by-domain refactoring - ready for merge
