+++ backend/src/main/java/com/veggieshop/auth/AuthService.java
@@ -1,6 +1,4 @@
-package com.veggieshop.service;
-
-import com.veggieshop.auth.AuthDto;
+package com.veggieshop.auth;
 
 public interface AuthService {
     AuthDto.AuthResponse login(AuthDto.AuthRequest request);
