+++ backend/src/main/java/com/veggieshop/category/CategoryRepository.java
@@ -1,6 +1,5 @@
-package com.veggieshop.repository;
+package com.veggieshop.category;
 
-import com.veggieshop.category.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
