+++ backend/src/main/java/com/veggieshop/offer/OfferDto.java
@@ -1,4 +1,4 @@
-package com.veggieshop.dto;
+package com.veggieshop.offer;
 
 import lombok.Data;
