+++ backend/src/main/java/com/veggieshop/offer/OfferRepository.java
@@ -1,6 +1,5 @@
-package com.veggieshop.repository;
+package com.veggieshop.offer;
 
-import com.veggieshop.offer.Offer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
