+++ backend/src/main/java/com/veggieshop/security/CurrentUserService.java
@@ -1,7 +1,7 @@
 package com.veggieshop.security;
 
-import com.veggieshop.entity.User;
-import com.veggieshop.repository.UserRepository;
+import com.veggieshop.user.User;
+import com.veggieshop.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
