+++ backend/src/main/java/com/veggieshop/security/UserDetailsServiceImpl.java
@@ -1,7 +1,6 @@
 package com.veggieshop.security;
 
-import com.veggieshop.entity.User;
-import com.veggieshop.repository.UserRepository;
+import com.veggieshop.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
